Add boolean resource isDefault=true mechanism
When given a location (e.g. by reading the SIM card network), AOSP determines the default keyboard based on the current location and this value.
For starters, create a proof of concept by only using the greek networks. Right now, we can't use it for all our languages, because our Keyboard needs to be able to switch to english during the SetupWizard (bug).